mv tune.syncer -> train._internal.syncer
