+++ python/ray/tune/execution/experiment_state.py
@@ -14,7 +14,7 @@ from ray.air._internal.uri_utils import _join_path_or_uri
 from ray.train._internal.storage import _use_storage_context, StorageContext
 from ray.tune.experiment import Trial
 from ray.tune.impl.out_of_band_serialize_dataset import out_of_band_serialize_dataset
-from ray.tune.syncer import SyncConfig, get_node_to_storage_syncer
+from ray.train._internal.syncer import SyncConfig, get_node_to_storage_syncer
 
 logger = logging.getLogger(__name__)
